% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdss_attribution_function.R
\name{cdss_attribution_function_sa}
\alias{cdss_attribution_function_sa}
\title{Build an attribution function from CDSS skill assignment matrices.}
\usage{
cdss_attribution_function_sa(sa)
}
\arguments{
\item{sa}{Skill assignment}
}
\value{
Data frame defining the attribution function derived from \code{sa}
}
\description{
\code{cdss_attribution_function_sa} builds an attribution function from
skill assignment matrices using the procedure described by Hockemeyer (2022).
}
