% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdss_attribution_function.R
\name{cdss_attribution_function_sma}
\alias{cdss_attribution_function_sma}
\title{Build an attribution function from CDSS skill multi-assignment matrices.}
\usage{
cdss_attribution_function_sma(sma)
}
\arguments{
\item{sma}{Skill multi-assignment}
}
\value{
Data frame defining the attribution function derived from \code{sma}
}
\description{
\code{cdss_attribution_function_sma} builds an attribution function from
skill multi-assignment matrices using the procedure described by Hockemeyer (2022).
}
