% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdss_binary_matrix_product.R
\name{cdss_binary_matrix_product}
\alias{cdss_binary_matrix_product}
\title{Compute a binary matrix product}
\usage{
cdss_binary_matrix_product(m, n)
}
\arguments{
\item{m}{Binary matrix}

\item{n}{Binary matrix}
}
\value{
Boolean matrix product of m and n
}
\description{
\code{cdss_binary_matrix_product} expects two binary matrices and computes there Boolean product.
}
\seealso{
Other Utility functions: 
\code{\link{cdss_close_ar}()},
\code{\link{cdss_reduce_sf}()}
}
\concept{Utility functions}
