% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdss_close_ar.R
\name{cdss_close_ar}
\alias{cdss_close_ar}
\title{Close an attribution relation to get a surmise relation.}
\usage{
cdss_close_ar(ar)
}
\arguments{
\item{ar}{Attribution relation matrix}
}
\value{
surmise relation or NULL
}
\description{
\code{cdss_close_ar} expects a quadratic binary matrix and closes it under
reflexivity and transitivity.
}
\seealso{
Other Utility functions: 
\code{\link{cdss_binary_matrix_product}()},
\code{\link{cdss_reduce_sf}()}
}
\concept{Utility functions}
