% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdss_csma2sf.R
\name{cdss_csma2sf}
\alias{cdss_csma2sf}
\title{Derive a surmise function from a complete skill multi-assignment}
\usage{
cdss_csma2sf(csma)
}
\arguments{
\item{csma}{Skill multi-assignment to be completed}
}
\value{
Object of class \code{cdss_csma}.
}
\description{
\code{cdss_csma2sf} expects a complete skill multi-assignment object 
and returns the corresponding surmise function on the set of skills.
}
