% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdss_lo_csma2sf.R
\name{cdss_lo_csma2sf}
\alias{cdss_lo_csma2sf}
\title{Derive a surmise function between learning objects from a complete skill multi-assignment}
\usage{
cdss_lo_csma2sf(csma)
}
\arguments{
\item{csma}{Complete skill multi-assignment object}
}
\value{
Object of class \code{cdss_sf} (attribution function).
}
\description{
\code{cdss_lo_csma2sf} expects a complete skill multi-assignment and derives a surmise
function on the set of learning objects.
}
\seealso{
Other functions building skill (multi) assignment matrices: 
\code{\link{cdss_lo_sa2af}()},
\code{\link{cdss_sa2sma}()},
\code{\link{cdss_tables2sa}()}
}
\concept{functions building skill (multi) assignment matrices}
