% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdss_lo_sa2af.R
\name{cdss_lo_sa2af}
\alias{cdss_lo_sa2af}
\title{Determine Attribution function for LOs from a skill assignment}
\usage{
cdss_lo_sa2af(sa)
}
\arguments{
\item{sa}{Skill assignment object}
}
\value{
Object of class \code{cdss_af} (attribution function).
}
\description{
\code{cdss_lo_sa2af} expects a skill assignment and derives an attribution
function on the set of learning objects.
}
\seealso{
Other functions building skill (multi) assignment matrices: 
\code{\link{cdss_lo_csma2sf}()},
\code{\link{cdss_sa2sma}()},
\code{\link{cdss_tables2sa}()}
}
\concept{functions building skill (multi) assignment matrices}
