% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdss_lo_sa2ar.R
\name{cdss_lo_sa2ar}
\alias{cdss_lo_sa2ar}
\title{Create an attribution relation on learning objects from a skill assignment.}
\usage{
cdss_lo_sa2ar(sa)
}
\arguments{
\item{sa}{Skill assignment object}
}
\value{
attribution relation or NULL
}
\description{
\code{cdss_lo_sa2ar} expects a skill assignment and derives an attribution relation
 on learning objects if the skill assignment fulfills the necessary conditions, i.e. if  
 there is only one teaching LO per skill.
}
\seealso{
Other functions deriving skill structures from skill assignments: 
\code{\link{cdss_sa2ar_skill}()},
\code{\link{cdss_sa_describes_sr}()}
}
\concept{functions deriving skill structures from skill assignments}
