% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdss_missing_los.R
\name{cdss_missing_los}
\alias{cdss_missing_los}
\title{Vector of skills without teaching learning objects.}
\usage{
cdss_missing_los(sa)
}
\arguments{
\item{sa}{SKill assignment}
}
\value{
Vector of skills
}
\description{
\code{cdss_missing_los} expects a skill assignment and returns a vector
of skills which are not taught by any learning object.
}
\seealso{
Other Functions testing validity of skill assignments: 
\code{\link{cdss_circular_requirements}()},
\code{\link{cdss_sa_compliance}()}
}
\concept{Functions testing validity of skill assignments}
