% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdss_nonteaching_los.R
\name{cdss_nonteaching_los}
\alias{cdss_nonteaching_los}
\title{Vector of learning objects not teaching any skills.}
\usage{
cdss_nonteaching_los(sa)
}
\arguments{
\item{sa}{SKill assignment}
}
\value{
Vector of learning objects
}
\description{
\code{cdss_nonteaching_los} expects a skill assignment and returns a vector
of learning objects which do not teach any skill.
}
\seealso{
Other Functions testing validity of skill assignments: 
\code{\link{cdss_circular_requirements}()},
\code{\link{cdss_missing_los}()},
\code{\link{cdss_sa_compliance}()}
}
\concept{Functions testing validity of skill assignments}
