% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdss_reduce_sf.R
\name{cdss_reduce_sf}
\alias{cdss_reduce_sf}
\title{Reduce a surmise function with respect to item equivalence}
\usage{
cdss_reduce_sf(sf)
}
\arguments{
\item{sf}{Surmise function}
}
\value{
Surmise function reduced by equivalences
}
\description{
\code{cdss_reduce_sf} takes a surmise function and returns its
reduction to non-equivalent items.
}
\seealso{
Other Utility functions: 
\code{\link{cdss_binary_matrix_product}()},
\code{\link{cdss_close_ar}()}
}
\concept{Utility functions}
