% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdss_sa2sma.R
\name{cdss_sa2sma}
\alias{cdss_sa2sma}
\title{Convert skill assignment matrices to skill multi-assignment}
\usage{
cdss_sa2sma(sa)
}
\arguments{
\item{sa}{Skill assignment object}
}
\value{
Object of class \code{cdss_sma}.
}
\description{
\code{cdss_sa2sma} expects a list of two matrices (\code{taught} and \code{required}) of a skill
assignment. It returns a skill multi-assignment object.
}
\seealso{
Other functions building skill (multi) assignment matrices: 
\code{\link{cdss_tables2sa}()}
}
\concept{functions building skill (multi) assignment matrices}
