% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdss_sa_compliance.R
\name{cdss_sa_compliance}
\alias{cdss_sa_compliance}
\title{Check whether a skill assignment is compliant to the CDCS conditions.}
\usage{
cdss_sa_compliance(sa, warnings = FALSE)
}
\arguments{
\item{sa}{Skill assignment}

\item{warnings}{Toggles whether warnings should be printed}
}
\value{
Boolean
}
\description{
\code{cdss_sa_compliance} expects a skill assignment and checks whether
it is compliant to the conditions for CDCS.
}
\seealso{
Other Functions testing validity of skill assignments: 
\code{\link{cdss_circular_requirements}()},
\code{\link{cdss_missing_los}()},
\code{\link{cdss_nonteaching_los}()}
}
\concept{Functions testing validity of skill assignments}
