% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdss_sa_describes_sr.R
\name{cdss_sa_describes_sr}
\alias{cdss_sa_describes_sr}
\title{Check whether a surmise relation can be derived from a given skill assignment.}
\usage{
cdss_sa_describes_sr(sa, verbose = FALSE)
}
\arguments{
\item{sa}{Skill assignment object}

\item{verbose}{Flag, default is FALSE}
}
\value{
Logical value
}
\description{
\code{cdss_sa_describes_sr} expects a list of two matrices (\code{taught} and \code{required}) of a skill
assignment. It returns TRUE if the skill assignment describes a surmise relation (i.e. there is
only one teaching LO per skill) and FALSE.
}
\seealso{
Other functions deriving skill structures from skill assignments: 
\code{\link{cdss_sa2ar_skill}()}
}
\concept{functions deriving skill structures from skill assignments}
