% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdss_sf2basis.R
\name{cdss_sf2basis}
\alias{cdss_sf2basis}
\title{Derive a basis from a surmise function}
\usage{
cdss_sf2basis(sf)
}
\arguments{
\item{sf}{Surmise function}
}
\value{
Matrix representing the basis.
}
\description{
\code{cdss_sf2basis} expects a surmise function object 
and returns the corresponding basis.
}
