% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdss_sma2csma.R
\name{cdss_sma2csma}
\alias{cdss_sma2csma}
\title{Complete a skill multi-assignment}
\usage{
cdss_sma2csma(sma)
}
\arguments{
\item{sma}{Skill multi-assignment to be completed}
}
\value{
Object of class \code{cdss_csma}.
}
\description{
\code{cdss_sma2csma} expects a skill multi-assignment object and returns
its closure under completeness.
}
