% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdss_tables2sa.R
\name{cdss_tables2sa}
\alias{cdss_tables2sa}
\title{Build matrices of taught and required, respectively, skills for learning objects from
respective tables.}
\usage{
cdss_tables2sa(taught, required)
}
\arguments{
\item{taught}{Data table containing the assignment of taught skills to learning objects}

\item{required}{Data table containing the assignment of required skills to learning objects}
}
\value{
List of two binary matrices, "taught" and "required".
}
\description{
\code{cdss_tables2sa} expects two data frames with two columns each. The first
column contains the IDs of learning objects and the second row the IDs of single skills
required or taught, respectively, by this learning object.
It returns a list of two binary matrices, "taught" and "required". Each matrix has one
row per learning object and one column per skill. The cells contain a "1" if the skill
is taught or required, respectively, by the learning object and a "0" otherwise.
}
\seealso{
Other functions building skill (multi) assignment matrices: 
\code{\link{cdss_lo_csma2sf}()},
\code{\link{cdss_lo_sa2af}()},
\code{\link{cdss_sa2sma}()}
}
\concept{functions building skill (multi) assignment matrices}
