% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close_ar.R
\name{close_ar}
\alias{close_ar}
\title{Close an attribution relation to get a surmise relation.}
\usage{
close_ar(ar)
}
\arguments{
\item{ar}{Attribution relation matrix}
}
\value{
surmise relation or NULL
}
\description{
\code{close_ar} expects a quadratic binary matrix and closes it under
reflexivity and transitivity.
}
\seealso{
Other Utility functions: 
\code{\link{binary_matrix_product}()},
\code{\link{reduce_sf}()}
}
\concept{Utility functions}
