% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_sf.R
\name{reduce_sf}
\alias{reduce_sf}
\title{Reduce a surmise function with respect to item equivalence}
\usage{
reduce_sf(sf)
}
\arguments{
\item{sf}{Surmise function}
}
\value{
Surmise function reduced by equivalences
}
\description{
\code{reduce_sf} takes a surmise function and returns its
reduction to non-equivalent items.
}
\seealso{
Other Utility functions: 
\code{\link{binary_matrix_product}()},
\code{\link{close_ar}()}
}
\concept{Utility functions}
